requirements.txt
fastapi==0.78.0
keras==2.7.0
numpy==1.21.4
opencv_python==4.5.5.62
pydantic==1.9.0
uvicorn==0.17.6
tensorflow-cpu==2.7.0
gunicorn==20.1.0